Drop redundant numeric casts across the tree
Remove the casts the redundant-scalar-cast check proves value-preserving:
int/float promotions an enclosing floating operation already performs,
float-to-double for non-overloaded double parameters and variadic
arguments, and same-width intermediate casts.

Whole-tree comparison: two bodies move slightly (W8AssayDialog::
CreateTextBuffers -2.3%, W8Quaternion::SetFromMatrix -0.8%), two
improve slightly; the SR model_instance TU is unchanged.

+++ include/surrender/srModelInstance.h
@@ -73,8 +73,8 @@ public:
         align_axis_14c = axis;
         length_squared = align_axis_14c.z * align_axis_14c.z + align_axis_14c.y * align_axis_14c.y +
                          align_axis_14c.x * align_axis_14c.x;
-        if ((double)length_squared != 0.0) {
-            scale = (float)(1.0 / sqrt((double)length_squared));
+        if (length_squared != 0.0) {
+            scale = (float)(1.0 / sqrt(length_squared));
             align_axis_14c *= scale;
         }
         alignment_flags_148.value |= 1;
